workspaces - shuffle some code around
