+++ src/vs/platform/windows/test/electron-main/window.test.ts
@@ -7,8 +7,7 @@ import * as assert from 'assert';
 import * as path from 'vs/base/common/path';
 import { findWindowOnFile } from 'vs/platform/windows/electron-main/windowsFinder';
 import { ICodeWindow, IWindowState } from 'vs/platform/windows/electron-main/windows';
-import { IWorkspaceIdentifier } from 'vs/platform/workspaces/common/workspaces';
-import { toWorkspaceFolders } from 'vs/platform/workspace/common/workspace';
+import { IWorkspaceIdentifier, toWorkspaceFolders } from 'vs/platform/workspaces/common/workspaces';
 import { URI } from 'vs/base/common/uri';
 import { getPathFromAmdModule } from 'vs/base/common/amd';
 import { extUriBiasedIgnorePathCase } from 'vs/base/common/resources';
