+++ src/vs/platform/workspace/common/workspace.ts
@@ -4,16 +4,17 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { URI } from 'vs/base/common/uri';
-import * as resources from 'vs/base/common/resources';
+import { joinPath, basenameOrAuthority } from 'vs/base/common/resources';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { TernarySearchTree } from 'vs/base/common/map';
 import { Event } from 'vs/base/common/event';
-import { IWorkspaceIdentifier, IStoredWorkspaceFolder, isRawFileWorkspaceFolder, isRawUriWorkspaceFolder, ISingleFolderWorkspaceIdentifier } from 'vs/platform/workspaces/common/workspaces';
+import { IWorkspaceIdentifier, IStoredWorkspaceFolder, ISingleFolderWorkspaceIdentifier } from 'vs/platform/workspaces/common/workspaces';
 import { IWorkspaceFolderProvider } from 'vs/base/common/labels';
 
 export const IWorkspaceContextService = createDecorator<IWorkspaceContextService>('contextService');
 
 export interface IWorkspaceContextService extends IWorkspaceFolderProvider {
+
 	readonly _serviceBrand: undefined;
 
 	/**
@@ -80,16 +81,6 @@ export interface IWorkspaceFoldersChangeEvent {
 	changed: IWorkspaceFolder[];
 }
 
-export namespace IWorkspace {
-	export function isIWorkspace(thing: unknown): thing is IWorkspace {
-		const candidate = thing as IWorkspace | undefined;
-
-		return !!(candidate && typeof candidate === 'object'
-			&& typeof candidate.id === 'string'
-			&& Array.isArray(candidate.folders));
-	}
-}
-
 export interface IWorkspace {
 
 	/**
@@ -106,6 +97,14 @@ export interface IWorkspace {
 	 * the location of the workspace configuration
 	 */
 	readonly configuration?: URI | null;
+}
+
+export function isWorkspace(thing: unknown): thing is IWorkspace {
+	const candidate = thing as IWorkspace | undefined;
+
+	return !!(candidate && typeof candidate === 'object'
+		&& typeof candidate.id === 'string'
+		&& Array.isArray(candidate.folders));
 }
 
 export interface IWorkspaceFolderData {
@@ -127,23 +126,21 @@ export interface IWorkspaceFolderData {
 	readonly index: number;
 }
 
-export namespace IWorkspaceFolder {
-	export function isIWorkspaceFolder(thing: unknown): thing is IWorkspaceFolder {
-		const candidate = thing as IWorkspaceFolder;
-
-		return !!(candidate && typeof candidate === 'object'
-			&& URI.isUri(candidate.uri)
-			&& typeof candidate.name === 'string'
-			&& typeof candidate.toResource === 'function');
-	}
-}
-
 export interface IWorkspaceFolder extends IWorkspaceFolderData {
 
 	/**
 	 * Given workspace folder relative path, returns the resource with the absolute path.
 	 */
 	toResource: (relativePath: string) => URI;
+}
+
+export function isWorkspaceFolder(thing: unknown): thing is IWorkspaceFolder {
+	const candidate = thing as IWorkspaceFolder;
+
+	return !!(candidate && typeof candidate === 'object'
+		&& URI.isUri(candidate.uri)
+		&& typeof candidate.name === 'string'
+		&& typeof candidate.toResource === 'function');
 }
 
 export class Workspace implements IWorkspace {
@@ -226,7 +223,7 @@ export class WorkspaceFolder implements IWorkspaceFolder {
 	}
 
 	toResource(relativePath: string): URI {
-		return resources.joinPath(this.uri, relativePath);
+		return joinPath(this.uri, relativePath);
 	}
 
 	toJSON(): IWorkspaceFolderData {
@@ -235,43 +232,5 @@ export class WorkspaceFolder implements IWorkspaceFolder {
 }
 
 export function toWorkspaceFolder(resource: URI): WorkspaceFolder {
-	return new WorkspaceFolder({ uri: resource, index: 0, name: resources.basenameOrAuthority(resource) }, { uri: resource.toString() });
-}
-
-export function toWorkspaceFolders(configuredFolders: IStoredWorkspaceFolder[], workspaceConfigFile: URI, extUri: resources.IExtUri): WorkspaceFolder[] {
-	let result: WorkspaceFolder[] = [];
-	let seen: Set<string> = new Set();
-
-	const relativeTo = extUri.dirname(workspaceConfigFile);
-	for (let configuredFolder of configuredFolders) {
-		let uri: URI | null = null;
-		if (isRawFileWorkspaceFolder(configuredFolder)) {
-			if (configuredFolder.path) {
-				uri = extUri.resolvePath(relativeTo, configuredFolder.path);
-			}
-		} else if (isRawUriWorkspaceFolder(configuredFolder)) {
-			try {
-				uri = URI.parse(configuredFolder.uri);
-				// this makes sure all workspace folder are absolute
-				if (uri.path[0] !== '/') {
-					uri = uri.with({ path: '/' + uri.path });
-				}
-			} catch (e) {
-				console.warn(e);
-				// ignore
-			}
-		}
-		if (uri) {
-			// remove duplicates
-			let comparisonKey = extUri.getComparisonKey(uri);
-			if (!seen.has(comparisonKey)) {
-				seen.add(comparisonKey);
-
-				const name = configuredFolder.name || extUri.basenameOrAuthority(uri);
-				result.push(new WorkspaceFolder({ uri, name, index: result.length }, configuredFolder));
-			}
-		}
-	}
-
-	return result;
+	return new WorkspaceFolder({ uri: resource, index: 0, name: basenameOrAuthority(resource) }, { uri: resource.toString() });
 }
